refactored shared_widgets
